+++ app/views/layouts/_nav_menu.html.erb
@@ -33,3 +33,4 @@
     </li>
   </ul>
 </div>
+<%= javascript_packs_with_chunks_tag "applyApplicationPolicyToggles", defer: true %>.
